+++ ImageFeed/AppDelegate.swift
@@ -9,9 +9,7 @@ import UIKit
 import ProgressHUD
 
 @main
-class AppDelegate: UIResponder, UIApplicationDelegate {
-
-
+final class AppDelegate: UIResponder, UIApplicationDelegate {
 
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
         ProgressHUD.animationType = .activityIndicator
@@ -40,7 +38,5 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         // If any sessions were discarded while the application was not running, this will be called shortly after application:didFinishLaunchingWithOptions.
         // Use this method to release any resources that were specific to the discarded scenes, as they will not return.
     }
-
-
 }
