+++ ImageFeed/TabBar/ImagesList/ImagesListService.swift
@@ -7,65 +7,28 @@
 
 import UIKit
 
-struct Photo {
-    let id: String
-    let size: CGSize
-    let createdAt: Date?
-    let welcomeDescription: String?
-    let largeImageURL: String
-    let thumbImageURL: String
-    var isLiked: Bool
-}
-struct PhotoLikeResult: Codable {
-    let photoResult: PhotoResult
-    
-    private enum CodingKeys: String, CodingKey {
-        case photoResult = "photo"
-    }
-}
-struct PhotoResult: Codable {
-    let id: String
-    let createdAt: String?
-    let width: Int
-    let height: Int
-    let isLiked: Bool
-    let description: String?
-    let images: UrlsResult
-    
-    private enum CodingKeys: String, CodingKey {
-        case id = "id"
-        case createdAt = "created_at"
-        case width = "width"
-        case height = "height"
-        case isLiked = "liked_by_user"
-        case description = "description"
-        case images = "urls"
-    }
-}
-
-struct UrlsResult: Codable {
-    let full: URL
-    let thumb: URL
-    
-    private enum CodingKeys: String, CodingKey {
-        case full = "full"
-        case thumb = "thumb"
-    }
-}
-
-enum ImageListServiceError: Error {
-    case invalidRequest
+private enum HTTPMethod: String {
+    case post = "POST"
+    case delete = "DELETE"
 }
 
 final class ImagesListService {
+    
+    // MARK: - Static Properties
     static let shared = ImagesListService()
     static let didChangeNotification = Notification.Name("ImagesListServiceDidChange")
     
+    // MARK: - Initializer
     private init() {}
     
+    // MARK: - Public Properties
     var photos: [Photo] = []
     
+    // MARK: - Private Properties
     private var lastLoadedPage: Int?
+    
+    let isoFormatter = ISO8601DateFormatter()
+    let outputFormatter = DateFormatter()
     
     private let token = OAuth2TokenStorage.shared.token
     private let urlSession = URLSession.shared
@@ -119,56 +82,6 @@ final class ImagesListService {
         }
     }
     
-    private func makeImagesListRequest(with page: Int) -> URLRequest? {
-        guard let apiURL = Constants.apiURL else {
-            print("APP: [ImageList] Unable to get base URL")
-            return nil
-        }
-        
-        guard let url = URL(string: "/photos?page=\(page)", relativeTo: apiURL) else {
-            print("APP: [ImageList] Unable to get photos image URL")
-            return nil
-        }
-        
-        var request = URLRequest(url: url)
-        guard let token else {
-            print("APP: Unable to get Bearer Token")
-            return nil
-        }
-        request.setValue("Bearer \(token)", forHTTPHeaderField: "Authorization")
-        
-        return request
-    }
-    
-    private func makePhotoFromRequest(_ photoUnsplash: PhotoResult) -> Photo {
-        var date = Date()
-        let isoFormatter = ISO8601DateFormatter()
-        isoFormatter.formatOptions = [
-            .withInternetDateTime,
-            .withDashSeparatorInDate,
-            .withColonSeparatorInTime
-        ]
-        if let photoDate = isoFormatter.date(from: photoUnsplash.createdAt!) {
-            let outputFormatter = DateFormatter()
-            outputFormatter.dateFormat = "d MMMM yyyy"
-            
-            let formattedDate = outputFormatter.string(from: photoDate)
-            print("Date: " + formattedDate)
-            date = photoDate
-        } else {
-            print("Date: Не удалось распознать дату")
-        }
-        return Photo(
-            id: photoUnsplash.id,
-            size: CGSize(width: photoUnsplash.width, height: photoUnsplash.height),
-            createdAt: date,
-            welcomeDescription: photoUnsplash.description ?? "",
-            largeImageURL: photoUnsplash.images.full.absoluteString,
-            thumbImageURL: photoUnsplash.images.thumb.absoluteString,
-            isLiked: photoUnsplash.isLiked
-        )
-    }
-    
     // MARK: - fetchLikeOrUnlikePhoto
     func changeLike(photoId: String, isLike: Bool, completion: @escaping (Result<Void, Error>) -> Void) {
         assert(Thread.isMainThread)
@@ -200,6 +113,55 @@ final class ImagesListService {
         task.resume()
     }
     
+    // MARK: - Requests
+    private func makeImagesListRequest(with page: Int) -> URLRequest? {
+        guard let apiURL = Constants.apiURL else {
+            print("APP: [ImageList] Unable to get base URL")
+            return nil
+        }
+        
+        guard let url = URL(string: "/photos?page=\(page)", relativeTo: apiURL) else {
+            print("APP: [ImageList] Unable to get photos image URL")
+            return nil
+        }
+        
+        var request = URLRequest(url: url)
+        guard let token else {
+            print("APP: Unable to get Bearer Token")
+            return nil
+        }
+        request.setValue("Bearer \(token)", forHTTPHeaderField: "Authorization")
+        
+        return request
+    }
+    
+    private func makePhotoFromRequest(_ photoUnsplash: PhotoResult) -> Photo {
+        var date = Date()
+        isoFormatter.formatOptions = [
+            .withInternetDateTime,
+            .withDashSeparatorInDate,
+            .withColonSeparatorInTime
+        ]
+        if let photoDate = isoFormatter.date(from: photoUnsplash.createdAt ?? "") {
+            outputFormatter.dateFormat = "d MMMM yyyy"
+            
+            let formattedDate = outputFormatter.string(from: photoDate)
+            print("Date: " + formattedDate)
+            date = photoDate
+        } else {
+            print("Date: Не удалось распознать дату")
+        }
+        return Photo(
+            id: photoUnsplash.id,
+            size: CGSize(width: photoUnsplash.width, height: photoUnsplash.height),
+            createdAt: date,
+            welcomeDescription: photoUnsplash.description ?? "",
+            largeImageURL: photoUnsplash.images.full.absoluteString,
+            thumbImageURL: photoUnsplash.images.thumb.absoluteString,
+            isLiked: photoUnsplash.isLiked
+        )
+    }
+    
     private func makeRequestLikeUnlikePhoto(id photoId: String, isLike: Bool) -> URLRequest? {
         guard let apiURL = Constants.apiURL else {
             print("APP: [ImageList] Unable to get base URL")
@@ -222,15 +184,11 @@ final class ImagesListService {
             return nil
         }
         request.setValue("Bearer \(token)", forHTTPHeaderField: "Authorization")
-        if isLike {
-            request.httpMethod = "POST"
-        } else {
-            request.httpMethod = "DELETE"
-        }
-        
+        request.httpMethod = isLike ? HTTPMethod.post.rawValue : HTTPMethod.delete.rawValue
         return request
     }
     
+    // MARK: - Private Methods
     private func updatePhotoLike(_ id: String, _ isLiked: Bool) {
         if let index = self.photos.firstIndex(where: { $0.id == id }) {
             photos[index].isLiked = isLiked
