+++ ImageFeed/TabBar/ImagesList/Models/ImageListServiceError.swift
@@ -1,0 +1,12 @@
+//
+//  ImageListServiceError.swift
+//  ImageFeed
+//
+//  Created by Александр Клопков on 24.06.2025.
+//
+
+import UIKit
+
+enum ImageListServiceError: Error {
+    case invalidRequest
+}
